addremove complete functionality
