--- src/modules/clear.js
@@ -1,9 +1,0 @@
-/** @format */
-// eslint-disable-next-line import/no-cycle
-import USER_INPUT from '../index.js';
-
-function clearInput() {
-  USER_INPUT.value = '';
-}
-
-export default clearInput;
