+++ src/modules/remove.js
@@ -1,6 +1,6 @@
+/* eslint-disable import/no-cycle */
 /** @format */
-import items from './get.js';
-import setItems from './set.js';
+import items, { setItems, refreshList } from './addremove.js';
 
 function removeItem() {
   const deleteButtons = document.querySelectorAll('.fa-solid.fa-trash');
@@ -8,7 +8,11 @@ function removeItem() {
     btn.addEventListener('click', (e) => {
       items.splice(id, 1);
       e.target.parentElement.remove();
+      items.forEach((item, index) => {
+        item.index = index + 1;
+      });
       setItems();
+      refreshList();
     });
   });
 }
