+++ src/style.css
@@ -12,7 +12,7 @@
   margin: 20px auto;
   border-radius: 5px;
   background: #fff;
-  box-shadow: 0 0 4px 5px rgba(0, 0, 0, 0.5);
+  box-shadow: 0 30px 30px 0 rgba(0, 0, 0, 0.5);
 }
 
 .header {
@@ -63,16 +63,15 @@ li {
   display: flex;
   align-items: center;
   gap: 30px;
-}
-
-.items li {
   justify-content: space-between;
   border-bottom: 1px solid rgb(212, 210, 210);
 }
 
 .item-description {
   font-size: 1.2rem;
-  color: black;
+  color: rgb(65, 62, 62);
+  border: none;
+  background-color: #fff;
 }
 
 h3 {
@@ -83,6 +82,10 @@ h3 {
 
 h3:hover {
   cursor: pointer;
+}
+
+.task {
+  padding: 15px 0;
 }
 
 .task i {
